fixed updater and other stuffs

+++ src/net/lapismc/HomeSpawn/LapisUpdater.java
@@ -10,7 +10,7 @@ import java.io.IOException;
 import java.net.URL;
 import java.nio.channels.Channels;
 import java.nio.channels.ReadableByteChannel;
-import java.util.Objects;
+import java.util.Date;
 
 public class LapisUpdater {
 
@@ -29,19 +29,27 @@ public class LapisUpdater {
             URL website = new URL("http://www.dart2112.comli.com/plugins/update.yml");
             ReadableByteChannel rbc = Channels.newChannel(website.openStream());
             File f = new File(plugin.getDataFolder().getAbsolutePath() + "update.yml");
-            FileOutputStream fos = new FileOutputStream(f);
-            fos.getChannel().transferFrom(rbc, 0, Long.MAX_VALUE);
-            rbc.close();
-            fos.flush();
-            fos.close();
-            YamlConfiguration yaml = YamlConfiguration.loadConfiguration(f);
-            String oldVersion = plugin.getDescription().getVersion();
-            String newVersion = yaml.getString(ID);
-            return !Objects.equals(oldVersion, newVersion);
+            Date d = new Date(f.lastModified());
+            Date d0 = new Date();
+            d0.setTime(d0.getTime() - 3600);
+            if (!f.exists() || d.before(d0)) {
+                FileOutputStream fos = new FileOutputStream(f);
+                fos.getChannel().transferFrom(rbc, 0, Long.MAX_VALUE);
+                rbc.close();
+                fos.flush();
+                fos.close();
+                YamlConfiguration yaml = YamlConfiguration.loadConfiguration(f);
+                String oldVersionString = plugin.getDescription().getVersion().replace(".", "");
+                String newVersionString = yaml.getString(ID).replace(".", "");
+                Integer oldVersion = Integer.parseInt(oldVersionString);
+                Integer newVersion = Integer.parseInt(newVersionString);
+                return oldVersion < newVersion;
+            } else {
+                return false;
+            }
         } catch (IOException e) {
             e.printStackTrace();
             return false;
         }
     }
-
 }
